requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-environ==0.11.2
drf-yasg==1.21.7
django-ratelimit==4.1.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0